requirements.txt
flask
tensorflow-cpu
numpy